Moved the ProcInfo js module into its own file so that we can share code between the old and new dashbaords.
