+++ linkerd/admin/src/main/resources/io/buoyant/linkerd/admin/template/process_info.template
@@ -1,5 +1,5 @@
 <div class="row text-center">
-  <div id="process-info" data-refresh-uri="/admin/metrics">
+  <div id="process-info">
     <ul class="list-inline topline-stats">
       {{#each stats}}
         <li data-key="{{dataKey}}">
